fix(files)!: second (and last) batch for rename files
