+++ src/models/artifactModel.ts
@@ -1,4 +1,4 @@
-import { CreateArtifactInput } from '../schemas/artifacts.js';
+import { CreateArtifactInput } from '../schemas/artifactsSchema.js';
 //Temporal, solo pruebas
 export interface Artifact extends CreateArtifactInput {
   id: string;
