+++ src/routes/artifactRoute.ts
@@ -1,11 +1,11 @@
 import { Router } from 'express';
 import { createArtifact, getArtifacts } from '../controllers/artifactController.js';
 import { verifyData } from '../middlewares/verifyData.js';
-import { createArtifactSchema } from '../schemas/artifacts.js';
+import { createArtifactSchema } from '../schemas/artifactsSchema.js';
 
 const router = Router();
 
-router.post("/artifacts",verifyData(createArtifactSchema),createArtifact);
+router.post("/artifacts", verifyData(createArtifactSchema), createArtifact);
 router.get("/artifacts", getArtifacts);
 
 export default router;
