+++ src/routes/roleRoute.ts
@@ -1,7 +1,7 @@
 import { Router } from "express"
 import { updateUserRole } from "../controllers/roleController.js";
 import { verifyData } from '../middlewares/verifyData.js'
-import { rolSchema } from '../schemas/role.js'
+import { rolSchema } from '../schemas/roleSchema.js'
 import { verifyRole, verifyToken } from "../middlewares/verifyToken.js"
 
 const router: Router = Router();
