+++ src/schemas/artifactsSchema.ts
